pyproject.toml
[tool.poetry]
name = "buckbotpy"
version = "0.1.0"
description = ""
authors = ["Curtis Babnik <curtisbabnik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
